% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle3.R
\name{run_fun}
\alias{run_fun}
\title{Run Monocle 3}
\usage{
run_fun(expression, parameters, priors, verbose, seed)
}
\arguments{
\item{expression}{Expression matrix}

\item{parameters}{Parameters}

\item{priors}{Priors}

\item{verbose}{Verbosity level}

\item{seed}{Random seed}
}
\description{
Run Monocle 3
}
\examples{
library(dyntoy)
dataset <- dyntoy::generate_dataset(
  num_cells = 99,
  num_features = 101,
  model = "tree",
  normalise = FALSE
)

expression <- dataset$expression
parameters <- list(
  num_dim = 50L,
  max_components = 2L,
  reduction_method = "UMAP",
  cluster_method = "leiden",
  k = 20L
)
priors <- dataset$prior_information
verbose <- TRUE
seed <- 1

output <- run_fun(expression, parameters, priors, verbose, seed)
}
